requirements.txt
Django==2.1.3
djangorestframework==3.9.0
pylint==2.1.1
celery==4.2.1
psycopg2-binary==2.7.6.1
Pillow == 5.3.0
